catch different types of fish, some helper functions

+++ resources/vrp/cfg/item/fishing.lua
@@ -6,13 +6,32 @@ local function start_fishing(player)
 		{'amb@world_human_stand_fishing@base','base',4},
 		{'amb@world_human_stand_fishing@idle_a','idle_c',1}
 	}
-	vRPclient.notify(player,{"~g~ Fishing"})
-	vRPclient.attachProp(player,{'prop_fishing_rod_01',60309,0,0,0,0,0,0})
-	vRPclient.playAnim(player,{false,seq,false})
-  --2558.50512695313,6155.3330078125,161.854034423828
-	SetTimeout(25000,function()
-		vRPclient.notify(player,{"~g~ Caught Fish"})
-		vRPclient.deleteProp(player,{'prop_fishing_rod_01'})
+	vRPclient.isInWater(player,{},function(truth)
+		if truth then 
+			vRPclient.notify(player,{"~g~ Fishing"})
+			vRPclient.attachProp(player,{'prop_fishing_rod_01',60309,0,0,0,0,0,0})
+			vRPclient.playAnim(player,{false,seq,false})
+			SetTimeout(24000,function()
+				vRPclient.getDistanceFrom(player,{2558.50512695313,6155.3330078125,161.854034423828},function(distance)
+					local caught
+					if distance < 100 then
+						caught = "high_quality_fish"
+					else
+						local keyset = {}
+						for k,v in pairs(items) do
+							table.insert(keyset,k)
+						end
+						caught = keyset[math.random(1,#keyset)]
+					end
+					user_id = vRP.getUserId(player)
+					vRP.giveInventoryItem(user_id,caught,1)
+					vRPclient.notify(player,{"~g~ Caught " .. items[caught][1]})
+					vRPclient.deleteProp(player,{'prop_fishing_rod_01'})
+				end)	
+			end)
+		else
+			vRPclient.notify(player,{"~r~You must be standing in water to fish."})
+		end
 	end)
 end
 
@@ -25,6 +44,10 @@ choices["Fish"] = {function(player,choice)
   end
 end}
 
-items["fishing_rod"] = {"Fishing Rod","A simple fishing rod.",choices,0.1}
+--{name,description,choices,weight}
+items["fishing_rod"] = {"Fishing Rod","A simple fishing rod.",choices,1.0}
+items["high_quality_fish"] = {"High Quality Fish","The best fish money can buy.",{},.2}
+items["regular_fish"] = {"Fish","Regular quality fish.",{},.2}
+items["low_quality_fish"] = {"Low Quality Fish","Low quality fish. Eating this is not reccomended.",{},.2}
 
 return items
